Store the cache's epoch-ms columns as BIGINT, not INTEGER
Postgres INTEGER is int4, and both cache.fetched_at (Date.now()) and
cache.ttl_ms (up to 30 days, from the Hunter lookup) are far past its
2,147,483,647 ceiling. Every external call goes through cached(), so the
first cache write in a search failed with

  value "1786948201233" is out of range for type integer

and took the whole search with it. Only Postgres was affected: SQLite's
INTEGER is 64-bit, so the identical sibling schema is fine and local dev
never saw it.

The schema change covers new databases. Existing ones need the widening,
which rewrites the table under an ACCESS EXCLUSIVE lock, so the fixup is
guarded on information_schema and runs at most once instead of on every
boot.

+++ lib/db/schema.postgres.sql
@@ -122,8 +122,11 @@ CREATE INDEX IF NOT EXISTS idx_opens_send ON email_opens(send_id, opened_at);
 CREATE TABLE IF NOT EXISTS cache (
   key         TEXT PRIMARY KEY,
   value       TEXT NOT NULL,                 -- JSON
-  fetched_at  INTEGER NOT NULL,              -- epoch ms
-  ttl_ms      INTEGER NOT NULL
+  -- BIGINT, not INTEGER: Postgres INTEGER is int4 and epoch ms passed the
+  -- 2^31 ceiling in 1970. SQLite's INTEGER is 64-bit, which is why the sibling
+  -- schema can say INTEGER and still work.
+  fetched_at  BIGINT NOT NULL,               -- epoch ms
+  ttl_ms      BIGINT NOT NULL
 );
 
 -- Learned per-domain email format, e.g. utoronto.ca -> "{first}.{last}".
